get ready for pypi release
