+++ .github/workflows/publish.yml
@@ -32,16 +32,26 @@ jobs:
           push: true
           tags: itsnebulalol/permasigner:latest
 
-  #pypi:
-  #  runs-on: ubuntu-latest
-  #  steps:
-  #    - name: Checkout
-  #      uses: actions/checkout@v2
+  pypi:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
 
-  #    - name: Build and publish to PyPi
-  #      uses: JRubics/poetry-publish@v1.12
-  #      with:
-  #        pypi_token: ${{ secrets.PYPI_TOKEN }}
-  #        ignore_dev_requirements: "yes"
-
+      - name: Set up Python
+        uses: actions/setup-python@v1
+        with:
+          python-version: '3.9'
       
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install setuptools wheel twine
+
+      - name: Build and publish
+        env:
+          TWINE_USERNAME: __token__
+          TWINE_PASSWORD: ${{ secrets.PYPI_TOKEN }}
+        run: |
+          python setup.py sdist bdist_wheel
+          twine upload dist/*
