+++ data/entitlements.plist
@@ -2,27 +2,27 @@
 <!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
 <plist version="1.0">
 <dict>
-  <key>com.apple.security.iokit-user-client-class</key>
-  <array>
-    <string>IOUserClient</string>
-  </array>
-  <key>platform-application</key>
-  <true/>
-  <key>get-task-allow</key>
-  <true/>
-  <key>keychain-access-groups</key>
-  <array>
+    <key>com.apple.security.iokit-user-client-class</key>
+    <array>
+      <string>IOUserClient</string>
+    </array>
+    <key>platform-application</key>
+    <true/>
+    <key>get-task-allow</key>
+    <true/>
+    <key>keychain-access-groups</key>
+    <array>
+      <string>{APP_BUNDLE}</string>
+    </array>
+    <key>application-identifier</key>
     <string>{APP_BUNDLE}</string>
-  </array>
-  <key>application-identifier</key>
-  <string>{APP_BUNDLE}</string>
-  <key>aps-environment</key>
-  <string>production</string>
-  <key>com.apple.developer.usernotifications.time-sensitive</key>
-  <true/>
-  <key>com.apple.security.application-groups</key>
-  <array>
-    <string>group.{APP_BUNDLE}</string>
-  </array>
+    <key>aps-environment</key>
+    <string>production</string>
+    <key>com.apple.developer.usernotifications.time-sensitive</key>
+    <true/>
+    <key>com.apple.security.application-groups</key>
+    <array>
+      <string>group.{APP_BUNDLE}</string>
+    </array>
 </dict>
 </plist>
