+++ permasigner/utils/hash.py
@@ -1,7 +1,7 @@
 import os
 import hashlib
 import requests
-from utils.downloader import Ldid
+from permasigner.utils.downloader import Ldid
 
 
 class Hash:
